StockMarketApp-main/StockMarketApp-main/requirements.txt
aiohttp==3.9.5
aiosignal==1.3.1
alembic==1.13.1
aniso8601==9.0.1
annotated-types==0.7.0
appdirs==1.4.4
asgiref==3.8.1
asttokens==2.4.1
attrs==23.2.0
beautifulsoup4==4.12.3
blinker==1.8.2
bs4==0.0.2
cachetools==5.3.3
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
cloudpickle==3.0.0
cmdstanpy==1.2.4
comm==0.2.2
contourpy==1.2.1
cssselect==1.2.0
cycler==0.12.1
dataclasses-json==0.6.6
debugpy==1.8.1
decorator==5.1.1
Deprecated==1.2.14
distlib==0.3.8
Django==5.0.7
django-cors-headers==3.13.0
djangorestframework==3.14.0
docker==7.1.0
entrypoints==0.4
executing==2.0.1
fake-useragent==1.5.1
feedparser==6.0.11
filelock==3.14.0
Flask==3.0.3
Flask-SQLAlchemy==3.1.1
fonttools==4.51.0
frozendict==2.4.4
frozenlist==1.4.1
fsspec==2024.5.0
gitdb==4.0.11
GitPython==3.1.43
graphene==3.3
graphql-core==3.2.3
graphql-relay==3.2.0
greenlet==3.0.3
gunicorn==22.0.0
holidays==0.51
html5lib==1.1
huggingface-hub==0.23.2
idna==3.7
importlib-metadata==7.0.0
importlib_resources==6.4.0
ipykernel==6.29.4
ipython==8.24.0
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
jsonpatch==1.33
jsonpointer==2.4
jsonschema==4.22.0
jsonschema-specifications==2023.12.1
jupyter_client==8.6.1
jupyter_core==5.7.2
kiwisolver==1.4.5
langchain==0.2.1
langchain-community==0.2.1
langchain-core==0.2.3
langchain-text-splitters==0.2.0
langsmith==0.1.65
lxml==5.2.2
Mako==1.3.5
Markdown==3.6
MarkupSafe==2.1.5
marshmallow==3.21.2
matplotlib==3.9.0
matplotlib-inline==0.1.7
mlflow==2.13.0
msgpack==1.0.8
multidict==6.0.5
multitasking==0.0.11
mypy-extensions==1.0.0
nest-asyncio==1.6.0
numpy==1.26.4
opentelemetry-api==1.24.0
opentelemetry-sdk==1.24.0
opentelemetry-semantic-conventions==0.45b0
orjson==3.10.3
packaging==23.2
pandas==2.2.2
parse==1.20.2
patsy==0.5.6
peewee==3.17.6
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.1
prompt-toolkit==3.0.43
prophet==1.1.5
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pydantic==2.7.2
pydantic_core==2.18.3
pyee==11.1.0
Pygments==2.17.2
prometheus_client
django-prometheus
requests_html
pyparsing==3.1.2
pyppeteer==2.0.0
pyquery==2.0.0
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
pyzmq==26.0.2
querystring-parser==1.2.4
ray>=2.31.0
referencing==0.35.1
requests==2.32.2
requests-html==0.10.0
rpds-py==0.18.1
scikit-learn==1.5.0
scipy==1.13.1
sgmllib3k==1.0.0
six==1.16.0
smmap==5.0.1
soupsieve==2.5
SQLAlchemy==2.0.30
sqlparse==0.5.0
stack-data==0.6.3
stanio==0.5.0
statsmodels==0.14.2
tenacity==8.3.0
threadpoolctl==3.5.0
tornado==6.4
tqdm==4.66.4
traitlets==5.14.3
typing-inspect==0.9.0
typing_extensions==4.11.0
tzdata==2024.1
urllib3==1.26.19
virtualenv==20.26.2
w3lib==2.2.1
wcwidth==0.2.13
webencodings==0.5.1
websockets==10.4
Werkzeug==3.0.3
wrapt==1.16.0
yahoo-fin==0.8.9.1
yarl==1.9.4
yfinance==0.2.41
zipp==3.18.2
yahoo_fin